Program can now add track to the layout and display it.
Changed the canvas class fields.
Imagesize default is 16.
There is now also Mode mode;
3 new methods in canvas class,
2 to calculate coordinates.
1 to draw StraightTracks (Currently only draws straightH but will be able to draw others soon)

In the map class added a new createAddStraightTrack method, this combines the creating track and adding it to the map.

Added update() when offsets change and when the user clicks.
Also changed the implementation of the chooseStraightH slot.

Also added the documentation for new methods.