Added a feature to connect link tracks
Added a method to draw hints when connecting linked tracks.
Added a method to connect link tracks in the map class.
Changed the method name of  setOtherLinkTrack to setOtherLinkedTrack.
Also added removeOtherLinkedTrack to LinkedTrack.
Added a minimum linked track constant and 2 error message methods in the window class, for a low number of LinkedTrack and off number of LinkedTracks.
Also updated CHANGELOG.md